Move root files to accelerator directory
