+++ accelerator/apis.c
