+++ accelerator/common_core.c
