+++ tests/BUILD
@@ -7,7 +7,7 @@ cc_test(
     ],
     deps = [
         ":iota_api_mock",
-        "//:common_core",
+        "//accelerator:common_core",
     ],
 )
